Put filter set form components in separate files
under /ExplorerFilterSetForms subdirectory
